first Storage class ok, recursive traversal is still missing

+++ amr2ugrid/AmrLevels.h
@@ -122,7 +122,7 @@ namespace Amr2Ugrid
 	in.read( (char*) levelSize, sizeof(AmrCellSize)*nLevels );
       }
 
-      int nLevels;
+      size_t nLevels;
       LevelInfo* levelInfo;
       AmrCellSize* levelSize;
     };
